Save the dbus connection in the adapter instance
