+++ src/wakepy/dbus_adapters/jeepney.py
@@ -1,18 +1,17 @@
 from __future__ import annotations
 
 import typing
+from typing import cast
 
 from jeepney import DBusAddress, new_method_call
-from jeepney.io.blocking import open_dbus_connection
+from jeepney.io.blocking import DBusConnection, open_dbus_connection
 from jeepney.wrappers import unwrap_msg
 
 from wakepy.core import DBusAdapter, DBusMethodCall
 from wakepy.core.dbus import BusType
 
 if typing.TYPE_CHECKING:
-    from typing import Dict, Optional, Union
-
-    from jeepney.io.blocking import DBusConnection
+    from typing import Optional, Union
 
     TypeOfBus = Optional[Union[BusType, str]]
 
@@ -29,9 +28,6 @@ class JeepneyDBusAdapter(DBusAdapter):
     # timeout for dbus calls, in seconds
     timeout = 2
 
-    def __init__(self) -> None:
-        self._connections: Dict[TypeOfBus, DBusConnection] = dict()  # type: ignore[no-any-unimported]
-
     def process(self, call: DBusMethodCall) -> object:
         addr = DBusAddress(
             object_path=call.method.path,
@@ -46,24 +42,10 @@ class JeepneyDBusAdapter(DBusAdapter):
             body=call.args,
         )
 
-        connection = self._get_connection(call.method.bus)
+        connection = cast(DBusConnection, self._get_connection(call.method.bus))  # type: ignore[no-any-unimported]
         reply = connection.send_and_get_reply(msg, timeout=self.timeout)
         resp = unwrap_msg(reply)
         return resp
-
-    def _get_connection(  # type: ignore[no-any-unimported]
-        self, bus: TypeOfBus = BusType.SESSION
-    ) -> DBusConnection:
-        """Gets either a new connection or a cached one, if there is such.
-        Caching of connections is done on bus level."""
-
-        if bus in self._connections:
-            return self._connections[bus]
-
-        connection = self._create_new_connection(bus)
-
-        self._connections[bus] = connection
-        return connection
 
     def _create_new_connection(  # type: ignore[no-any-unimported]
         self, bus: TypeOfBus = BusType.SESSION
