+++ tests/integration/conftest.py
@@ -6,6 +6,7 @@ import gc
 import logging
 import subprocess
 import sys
+import warnings
 
 import pytest
 
@@ -32,7 +33,11 @@ def gc_collect_after_dbus_integration_tests():
     # this as garbage colletion is triggered also automatically. The garbage
     # collection must be triggered here manually as the warnings are
     # ResourceWarning is only filtered away in the dbus integration tests.
-    gc.collect()
+
+    with warnings.catch_warnings():
+        warnings.simplefilter("ignore")
+        gc.collect()
+
     logger.debug("called gc.collect")
